requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
vobject

